new end-point to get parent pub ID and source url
