+++ web-app/src/routes/api/posts/+server.ts
@@ -1,75 +1,46 @@
 import {PUBLIC_LENS_API_URL} from "$env/static/public";
-import getPosts from '../../../graphql/getPosts';
 import {json} from '@sveltejs/kit';
-import {APP_LENS_ID} from '$env/static/private';
 import getComments from "../../../graphql/getComments";
-
+import {getParentPost} from "../../../utils/backend/get-parent-url.server";
 
 export async function GET(request) {
-    let statusCode;
 
-    const hashedURL = request.url.searchParams.get('hashedURL');
     try {
-        const posts = await fetch(PUBLIC_LENS_API_URL, {
-                method: 'POST',
-                headers: {
-                    'Content-Type': 'application/json',
+        const res = await getParentPost(request);
+        const parentPostID = res['parent_post_ID'];
+        const sourceURL = res['source_url'];
+
+        const comments = await fetch(PUBLIC_LENS_API_URL, {
+            method: 'POST',
+            headers: {
+                'Content-Type': 'application/json',
+            },
+            body: JSON.stringify({
+                query: getComments,
+                variables: {
+                    publicationId: parentPostID
                 },
-                body: JSON.stringify({
-                    query: getPosts,
-                    variables: {
-                      hashedURL: hashedURL,
-                      lensId: APP_LENS_ID
-                    },
-                }),
-            })
-        statusCode = posts.status;
-        const response = await posts.json()
-        const parentPostID = response.data.publications.items[0].id;
-        const sourceURL = response.data.publications.items[0].metadata.description;
+            }),
+        });
 
-        try {
-            const comments = await fetch(PUBLIC_LENS_API_URL, {
-                method: 'POST',
-                headers: {
-                    'Content-Type': 'application/json',
-                },
-                body: JSON.stringify({
-                    query: getComments,
-                    variables: {
-                        publicationId: parentPostID
-                    },
-                }),
-            })
-            statusCode = comments.status;
+        const commentsJSON = await comments.json();
+        const commentItems = commentsJSON.data.publications.items;
 
-            const commentsJSON = await comments.json();
-            const commentItems = commentsJSON.data.publications.items;
-
-            const responseComments = {
-                status_code: statusCode,
-                URL: sourceURL,
-                parentPublicationID: parentPostID,
-                items: commentItems
-            };
+        const responseComments = {
+            status_code: 200,
+            URL: sourceURL,
+            parentPublicationID: parentPostID,
+            items: commentItems
+        };
 
 
-            return json(responseComments)
-
-        } catch (err) {
-            console.log(err);
-            return json({
-                status_code: statusCode,
-                error: `Could not fetch comments for Publication ID, hashed URl: ${parentPostID}`
-            });
-        }
+        return json(responseComments)
 
     } catch (err) {
         console.log(err);
         return json({
-            status_code: statusCode,
-            error: `Could not fetch Publication ID for hashed URl: ${hashedURL}`
+            status_code: 404,
+            error: `Could not fetch comments for Publication ID, hashed URl: ${request.url.searchParams.get('hashedURL')}`
         });
     }
 }
-
